Add patches to switch the deployment to cluster scope
It used to be the default that the kustomize-generated manifests
installed Marin3r using ClusterRoles so one manager pod could look for
envoydeployments cluster-wide. This changed at some point but there's
code in config/default/kustomization.yaml to put it back.

Instead of modifying upstream files I copied them into the config/epic
directory that I use to generate the deployment manifests that we use
in EPIC. This should hopefully be easier to carry along as a patch
kit.