Remove Liveness and Readiness probes
Not sure why, but with the config changes we make to our Envoy pods,
k8s isn't able to reach the admin endpoint so it kept restarting the
pods.

Fixing this will be one of the first jobs we need to do when we figure
out how to advertise routes only to Envoy pods that have valid
upstream endpoints.