Use the current image as the default for Envoy init and shutdown
While the Envoy init manager and shutdown manager images can be
overridden, the defaults were more-or-less hardcoded. Now the default
image is the same as what was built, which is usually correct.

NOTE: I wanted to use the image that the manager container is
currently running, but that turns out to be very difficult to
find. The image name isn't available in the downstream API so for now
I'm doing a cut-n-paste of the existing "version" code and having the
Makefile pass the image into the generator.