Switch the readiness probe to the data port
We want the readiness probe to test as much of the path from Envoy to
its upstreams as possible, but "/ready" on the admin port is just a
duplicate of the liveness probe, i.e., simply whether Envoy is
running.

This change requires that the upstream have a "/ready" endpoint (so we
might need to make this configurable), but it tests the readiness of
Envoy, TrueIngress and its tunnel config, and the upstreams.