Log normally-occurring "errors" only once
Resource updates can fail as part of normal operations, for example,
when two controllers try to update the same object at the same
time. This is OK. When the controller returns an error to the
controller manager the request will be retried. In the mean time,
though, we were throwing two ERROR level events into the log: one of
our own, and one logged by the controller manager. This adds a lot of
scary chaff to the logs when things are actually working.

This commit removes our calls to log.Error() for things that are
expected to happen, and don't cause any bad effects.