Don't get confused when re-publishing a previously-published version
The EC.Status desiredVersion and publishedVersion were getting
confused if you published one rev, then another, then re-published the
first one. You'd end up with the EC in a weird state where the desired
version and published version were different and there was a
CacheOutOfSync=True condition in the EC Status.Conditions.

The cause was that RevisionReconciler.Reconcile() assumed that the
ECR with the most recent LastPublishedAt was the one that we wanted to
publish, but we never updated that field so in this case it would be
older than the other ECR's LastPublishedAt.

This patch "touches" the desired ECR's LastPublishedAt which forces it
to be chosen by getVersionToPublish().