Correctly sort ECRs when there are only two in the list
SortByPublication had a bug where it was checking only one of the
candidate's version but not the other. If there are only two
candidates to be sorted then one's version wasn't being examined so
the sort ended up being based on publish date only, ignoring the
version.

Now we check both candidates' versions so version is always the
primary sort key, even when there are only two candidates.